+++ lib/application/helper/reviews.dart
@@ -4,6 +4,7 @@ import 'package:universal_io/io.dart';
 
 // Project imports:
 import 'package:da_kanji_mobile/globals.dart';
+import 'package:url_launcher/url_launcher_string.dart';
 
 /// Opens a in app review dialogue if available or otherwise
 /// opens the app in the platforms app store
@@ -13,11 +14,13 @@ void openReview() async {
   if((Platform.isAndroid || Platform.isIOS || Platform.isMacOS) &&
     await inAppReview.isAvailable()) {
     inAppReview.requestReview();
-  } else {
+  } else if(Platform.isWindows) {
     inAppReview.openStoreListing(
       appStoreId: g_AppStoreId, 
       microsoftStoreId: g_MicrosoftStoreId
     );
+  } else if(Platform.isLinux) {
+    launchUrlString(g_SnapStorePage);
   }
 
 }