+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Fixed:
 - Pressing "rate" on linux causes exception
 - Importing the dojg deck
 - DoJG deep links
+- DoJG screen search text color
 
 ## v3.4.1 - ワードリストと暗記改善 (09.06.2024)