+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 
 # DaKanji - changelog
 
+## v3.4.2 - ワードリストと暗記改善 (09.06.2024)
+
+Fixed:
+
+- Pressing "rate" on linux causes exception
+- Importing the dojg deck
+- DoJG deep links
+- DoJG screen search text color
+
 ## v3.4.1 - ワードリストと暗記改善 (09.06.2024)
 
 Added: